requirements_streamlit.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
requests
beautifulsoup4
pypdf
python-dotenv
